web/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "advanced-vuln-scanner"
version = "0.1.0"
description = "Prototype web vulnerability scanner focusing on OWASP Top 10 categories."
requires-python = ">=3.10"
authors = [
  {name = "Security Research Team"}
]
dependencies = [
    "httpx>=0.27",
    "rich>=13.7",
    "pydantic>=2.7",
    "pyyaml>=6.0.1",
    "beautifulsoup4>=4.12",
    "tenacity>=8.2",
    "flask>=3.0",
    "flask-cors>=4.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-asyncio>=0.23",
  "ruff>=0.6",
  "mypy>=1.10"
]

[project.scripts]
vuln-scanner = "scanner.main:app"

[tool.setuptools.packages.find]
include = ["scanner*"]
exclude = ["configs", "tests"]


